refactor: create new configService category
This commit refactors the new Settings route to use a new
SettingsService so that business logic in the routes layer is
simplified - the details of HOW the Settings files are updated
should not be exposed on the route layer.

In doing so, we created a new configServices directory to store
the SettingsService, since it does not exist as a file service.
